+++ src/main/java/scs/planus/global/auth/service/OAuthService.java
@@ -10,8 +10,8 @@ import scs.planus.global.auth.dto.OAuthLoginResponseDto;
 import scs.planus.global.auth.dto.apple.AppleAuthRequestDto;
 import scs.planus.global.auth.dto.apple.AppleClientSecretResponseDto;
 import scs.planus.global.auth.dto.apple.FullName;
-import scs.planus.global.auth.entity.AppleUserInfo;
-import scs.planus.global.auth.entity.OAuthUserInfo;
+import scs.planus.global.auth.entity.userinfo.AppleUserInfo;
+import scs.planus.global.auth.entity.userinfo.OAuthUserInfo;
 import scs.planus.global.auth.entity.Token;
 import scs.planus.global.auth.service.apple.AppleJwtProvider;
 import scs.planus.global.auth.service.apple.AppleOAuthUserProvider;
