+++ src/main/java/scs/planus/global/auth/service/apple/AppleAuthClient.java
@@ -3,7 +3,7 @@ package scs.planus.global.auth.service.apple;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
-import scs.planus.global.auth.entity.apple.ApplePublicKeys;
+import scs.planus.global.auth.entity.ApplePublicKeys;
 
 @Component
 public class AppleAuthClient {
