+++ src/main/java/scs/planus/global/auth/service/google/GoogleUserInfoClient.java
@@ -3,8 +3,8 @@ package scs.planus.global.auth.service.google;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
-import scs.planus.global.auth.entity.GoogleUserInfo;
-import scs.planus.global.auth.entity.OAuthUserInfo;
+import scs.planus.global.auth.entity.userinfo.GoogleUserInfo;
+import scs.planus.global.auth.entity.userinfo.OAuthUserInfo;
 
 @Component
 public class GoogleUserInfoClient {
