+++ src/main/java/scs/planus/global/auth/service/kakao/KakaoOAuthUserProvider.java
@@ -4,8 +4,8 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import scs.planus.global.auth.dto.OAuthTokenResponseDto;
-import scs.planus.global.auth.entity.KakaoUserInfo;
-import scs.planus.global.auth.entity.OAuthUserInfo;
+import scs.planus.global.auth.entity.userinfo.KakaoUserInfo;
+import scs.planus.global.auth.entity.userinfo.OAuthUserInfo;
 
 @Component
 @RequiredArgsConstructor
